+++ src/backend-project/src/main/java/io/swen90007sm2/app/model/entity/Room.java
@@ -16,7 +16,7 @@ public class Room extends BaseEntity {
 
     private String description;
 
-    private Double pricePerNight = 0.0;
+    private BigDecimal pricePerNight = BigDecimal.valueOf(0.0);
 
     private String currency = CommonConstant.AUD_CURRENCY;
 
@@ -33,7 +33,7 @@ public class Room extends BaseEntity {
         super(createTime, updateTime);
     }
 
-    public Room(String roomId, String hotelId, String name, String description, Double pricePerNight, String currency, Integer sleepsNum, Integer vacantNum, Boolean onSale) {
+    public Room(String roomId, String hotelId, String name, String description, BigDecimal pricePerNight, String currency, Integer sleepsNum, Integer vacantNum, Boolean onSale) {
         this.roomId = roomId;
         this.hotelId = hotelId;
         this.name = name;
@@ -45,7 +45,7 @@ public class Room extends BaseEntity {
         this.onSale = onSale;
     }
 
-    public Room(Date createTime, Date updateTime, String roomId, String hotelId, String name, String description, Double pricePerNight, String currency, Integer sleepsNum, Integer vacantNum, Boolean onSale) {
+    public Room(Date createTime, Date updateTime, String roomId, String hotelId, String name, String description, BigDecimal pricePerNight, String currency, Integer sleepsNum, Integer vacantNum, Boolean onSale) {
         super(createTime, updateTime);
         this.roomId = roomId;
         this.hotelId = hotelId;
@@ -58,7 +58,7 @@ public class Room extends BaseEntity {
         this.onSale = onSale;
     }
 
-    public Room(Date createTime, Date updateTime, String roomId, String hotelId, String name, String description, Double pricePerNight, Integer sleepsNum, Integer vacantNum, Boolean onSale) {
+    public Room(Date createTime, Date updateTime, String roomId, String hotelId, String name, String description, BigDecimal pricePerNight, Integer sleepsNum, Integer vacantNum, Boolean onSale) {
         super(createTime, updateTime);
         this.roomId = roomId;
         this.hotelId = hotelId;
@@ -94,11 +94,11 @@ public class Room extends BaseEntity {
         this.description = description;
     }
 
-    public Double getPricePerNight() {
+    public BigDecimal getPricePerNight() {
         return pricePerNight;
     }
 
-    public void setPricePerNight(Double pricePerNight) {
+    public void setPricePerNight(BigDecimal pricePerNight) {
         this.pricePerNight = pricePerNight;
     }
